Use middleware strategy to protect admin endpoints

+++ src/routes/web.php
@@ -47,13 +47,13 @@ Route::delete('/cars/delete/{id}', 'App\Http\Controllers\CarController@delete')-
 // Car model
 Route::get('/car-model', 'App\Http\Controllers\CarModelController@index')->name('carModel.index');
 
-Route::get('/car-model/create', 'App\Http\Controllers\CarModelController@create')->name('carModel.create')->middleware('auth');
+Route::get('/car-model/create', 'App\Http\Controllers\CarModelController@create')->name('carModel.create')->middleware('auth', 'isAdmin');
 
 Route::get('/car-model/{id}', 'App\Http\Controllers\CarModelController@show')->name('carModel.show');
 
-Route::post('/car-model/save', 'App\Http\Controllers\CarModelController@save')->name('carModel.save')->middleware('auth');
+Route::post('/car-model/save', 'App\Http\Controllers\CarModelController@save')->name('carModel.save')->middleware('auth', 'isAdmin');
 
-Route::delete('/car-model/delete/{id}', 'App\Http\Controllers\CarModelController@delete')->name('carModel.delete')->middleware('auth');
+Route::delete('/car-model/delete/{id}', 'App\Http\Controllers\CarModelController@delete')->name('carModel.delete')->middleware('auth', 'isAdmin');
 
 // Publish request
 Route::get('/publish-request', 'App\Http\Controllers\PublishRequestController@index')->name('publishRequest.index');
